2008-02-28  Rafael Espindola  <espindola@google.com>
	   * tree-dfa.c (dump_dfa_stats): cast dfa_stats.max_num_phi_args to long.



git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/gimple-tuples-branch@132746 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/tree-dfa.c
@@ -495,7 +495,7 @@ dump_dfa_stats (FILE *file)
   if (dfa_stats.num_phis)
     fprintf (file, "Average number of arguments per PHI node: %.1f (max: %ld)\n",
 	     (float) dfa_stats.num_phi_args / (float) dfa_stats.num_phis,
-	     dfa_stats.max_num_phi_args);
+	     (long) dfa_stats.max_num_phi_args);
 
   fprintf (file, "\n");
 }
